+++ addons/text_editor/TE_FileTabs.gd
@@ -1,6 +1,6 @@
 extends TabContainer
 
-onready var editor:TE_TextEditor = owner
+onready var editor:TE_Editor = owner
 var mouse:bool = false
 
 func _ready():