+++ addons/text_editor/ext/ext_yaml.gd
@@ -43,7 +43,7 @@ func toggle_comment(t:TextEdit, head:String="", tail:String=""):
 	
 	return [old, new]
 
-func apply_colors(e:TE_TextEditor, t:TextEdit):
+func apply_colors(e:TE_Editor, t:TextEdit):
 	.apply_colors(e, t)
 	
 	# strings